feature(interceptors): Take out empty interceptors into separate class;
---

Closes #15
