+++ tests/Interceptor/InterceptorAroundCallableTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Panelop\Core\Tests\Interceptor;
 
+use Interceptor\EmptyInterceptorTest;
 use Panelop\Core\Interceptor\Builders\InterceptorBuilder;
 use Panelop\Core\Interceptor\Factories\InvocationMethodFactory;
 use Panelop\Core\Interceptor\Interfaces\InterceptorAroundInterface;
@@ -12,11 +13,10 @@ use Panelop\Core\Interceptor\InvocationAroundResult;
 use Panelop\Core\Interceptor\InvocationMethod;
 use Panelop\Core\Interceptor\InvocationParameter;
 use PHPUnit\Framework\Attributes\CoversClass;
-use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\DataProviderExternal;
 use PHPUnit\Framework\TestCase;
 
 use function array_values;
-use function func_get_args;
 
 #[CoversClass(InterceptorAroundInterface::class)]
 #[CoversClass(InvocationMethodFactory::class)]
@@ -25,52 +25,10 @@ use function func_get_args;
 #[CoversClass(InterceptorBuilder::class)]
 final class InterceptorAroundCallableTest extends TestCase
 {
-    public static function dataProvider(): iterable
-    {
-        yield "Two Parameters Function" => [
-            'callable' => static fn (mixed $first, mixed $second): array => func_get_args(),
-            'hasVariadic' => false,
-        ];
-
-        yield "Two Parameters And Variadic Function" => [
-            'callable' => static fn (mixed $first, mixed $second, mixed ...$last): array => func_get_args(),
-            'hasVariadic' => true,
-        ];
-
-        yield "Variadic Function" => [
-            'callable' => static fn (mixed ...$variadic): array => func_get_args(),
-            'hasVariadic' => true,
-        ];
-    }
-
-    #[DataProvider('dataProvider')]
-    public function testEmptyInterceptors(
-        callable $callable,
-        bool $hasVariadic,
-    ): void {
-        self::assertSame(['Hello', 'world!'], $callable('Hello', 'world!'));
-        self::assertSame([1, 2], $callable(1, 2));
-
-        if ($hasVariadic) {
-            self::assertSame([1, 2, 3, 4, 5], $callable(1, 2, 3, 4, 5));
-        }
-
-        $interceptor = (new InterceptorBuilder())
-            ->on((new InvocationMethodFactory())->create($callable))
-            ->build();
-
-        self::assertSame(['Hello', 'world!'], $interceptor('Hello', 'world!'));
-        self::assertSame([1, 2], $interceptor(1, 2));
-
-        if ($hasVariadic) {
-            self::assertSame([1, 2, 3, 4, 5], $interceptor(1, 2, 3, 4, 5));
-        }
-    }
-
-    #[DataProvider('dataProvider')]
+    #[DataProviderExternal(EmptyInterceptorTest::class, 'dataProvider')]
     public function testCallable(
         callable $callable,
-        bool $hasVariadic,
+        bool     $hasVariadic,
     ): void {
         self::assertSame(['Hello', 'world!'], $callable('Hello', 'world!'));
         self::assertSame([1, 2], $callable(1, 2));
@@ -137,4 +95,6 @@ final class InterceptorAroundCallableTest extends TestCase
             self::assertSame([1, "-- it is the result --", 3, 4, 5], $interceptor(1, 2, 3, 4, 5));
         }
     }
+
+
 }
